// Giao hàng/types.ts
export interface Product {
  id: string;
  name: string;
  category: string; // 85g, Đã nem 20, Lá...
  unit: string; // ĐVT
  price: number;
  weightKg: number; // For load calculation
}